+++ smoderp2d/runners/grass.py
@@ -1,0 +1,8 @@
+from smoderp2d.runners.base import Runner
+                    
+class GrassGisRunner(Runner):
+    """TODO."""
+    def _get_provider(self):
+        from smoderp2d.providers.grass import GrassGisProvider
+        return GrassGisProvider()
+
